Tests: optionally prevent autospec helper from passing mocks
Not everything that's decorated needs the mocks that are patched. Being
required to add the args to the test function anyway is annoying. It's
especially bad if trying to decorate an entire test suite, as every
test would need the args.

Move the definition to a separate module to keep things cleaner.