Bot: add root alias support
Override `Bot.add_command` and `Bot.remove_command` to add/remove root
aliases for a command (and recursively for any subcommands). This has
to happen in `Bot` because there's no reliable way to get the `Bot`
instance otherwise. Therefore, overriding the methods in `GroupMixin`
unfortunately doesn't work. Otherwise, it'd be possible to avoid
recursion by processing each subcommand as it got added.