Verification: stop kicking members on suspicious 403
A Discord error code 50_0007 signifies that the DM dispatch failed
because the target user does not accept DMs from the bot.
Such errors are ignored as before.

Any other 403s will however cause the bot to stop making requests.
This is in case the bot gets caught by an anti-spam filter and
should immediately stop.