Disable burst_shared filter of the AntiSpam cog
Our AntiSpam cog suffers from a race condition that causes it to try and
infract the same user multiple times. As that happens frequently with
the burst_shared filter, it means that our bot joins in and starts
spamming the channel with error messages.

Another issue is that burst_shared may cause our bot to send a lot of
DMs to a lot of different members. This caused our bot to get a DM ban
from Discord after a recent `everyone` ping incident.

I've decided to disable the `burst_shared` filter by commenting out the
relevant lines but leave the code in place otherwise. This means we
still have the implementation handy in case we want to re-enable it on
short notice.

Signed-off-by: Sebastiaan Zeeff <sebastiaan.zeeff@gmail.com>