Use global namespace `bot` for our RedisSession
As we're now planning on using Redis in multiple applications, it's
important to minimize the risk of namespace collisions between different
applications.

The `async-rediscache` packages allows us to set a global namespace on
an application level. I've chosen "bot" as the namespace for this
application, which means all individual namespaces will automatically be
prefixed by `bot.` whenever they are accessed.

Signed-off-by: Sebastiaan Zeeff <sebastiaan.zeeff@gmail.com>