Determine eligible duckpond emojis dynamically
Instead of maintaining a list of duckpond IDs manually, it's a much
better idea to detect ducky emojis dynamically using the new emoji name
grouping we've introduced: All emojis that start with `ducky_` will now
be counted as a duckpond ducky.

The unicode duck emoji obviously still counts in addition to custom
emojis with the `ducky_` prefix.

Signed-off-by: Sebastiaan Zeeff <sebastiaan.zeeff@gmail.com>