Code block: fix _fix_indentation failing for line counts of 1
This could be reproduced by editing a tracked message to a single line
of invalid Python that lacks any back ticks. The code was assuming there
would be multiple lines because that's what the default value for the
threshold is, but this threshold is not applied to edited messages.

Fixes BOT-A5