Detect codeblock language with special characters
The regex we use to detect codeblocks did not recognize language
specifiers that use a dash, a plus, or a dot in their name. As there are
valid language specifiers, such as python-repl and c++, that use those
characters, I've changed the regex to reflect that.

The character set used now reflects the characters used in language
specifiers in highlight.js.

Signed-off-by: Sebastiaan Zeeff <sebastiaan.zeeff@gmail.com>