Voice Gate: define atomic `_delete_ping` function
The code for ping deletion was duplicated in two places. In this commit,
we move it into a helper function, and apply a lock to make each
transaction atomic.

This means that if two coroutines try to call the function, the first
has to finish before the second can begin. This avoids the following:

Coro1: Message in cache? Yes.
Coro1: Send delete request. Yield control (await).
Coro2: Message in cache? Yes.

Now Coro2 has to wait for Coro1 to finish. Therefore it will always
find the `NO_MSG` signal, and not attempt the deletion.

Co-authored-by: MarkKoz <KozlovMark@gmail.com>
Co-authored-by: Sebastiaan Zeeff <sebastiaan.zeeff@gmail.com>
Co-authored-by: Daniel Brown <browndj3@gmail.com>