Voice Gate: ensure atomicity when notifying users
Previously, the listener risked yielding control to a racing event at
multiple points between checking whether the member was already
notified, notifying them, and writing this information into the cache.

As a result, in a pathological case, multiple racing coroutines could
have passed the membership check and ping-spammed the user, before the
first coro could have a chance to write the message ID into the cache.

In this commit, we move this logic into an atomic helper, which will
ensure that events are processed one-by-one, and subsequent events
correctly abort.

Co-authored-by: MarkKoz <KozlovMark@gmail.com>
Co-authored-by: Sebastiaan Zeeff <sebastiaan.zeeff@gmail.com>
Co-authored-by: Daniel Brown <browndj3@gmail.com>