CI: disable 'continueOnError'
After #1219, we started to encounter issues with the cache being
corrupted and CI failing due to 'pre-commit' not being installed
after restore.

Although it doesn't seem likely that this could have been the culprit,
the issues began appearing shortly after merging the PR.

Let's see what happens if we disable it.