Migrate CI Pipeline to GitHub Actions
I've migrated our Azure CI Pipeline to GitHub Actions. While the general
workflow is the same, there are a few changes:

- `flake8` is no longer run by `pre-commit`, but rather by a separate
  action that adds annotations to the GH Action results page.

- As we no longer have need for xml-formatted coverage files, the
  xmlrunner for unittest has been removed as a dependency. Instead, we
  now publish our coverage results to coveralls.io.

- We use version 2 of docker's GitHub Action build-and-push flow, which
  is split over multiple steps instead of one.

- I have changed the badges to GitHub Actions and coveralls.io badges.

Note: Because we accept PRs from forks, we need to be a bit careful with
our secrets. While we do use the `pull_request_target` event, we should
not expose secrets in steps that run code from the repository.

Signed-off-by: Sebastiaan Zeeff <sebastiaan.zeeff@gmail.com>