Make sure we lint the actual pull request
Unfortunately, our old setup did not actually lint the PR, as it was
running in the context of the target repository. To sidestep the issue
of using `pull_request_target` altogether, I've now changed our run of
flake8 to using it directly and having it output its errors in teh
format of Workflow Commands.

This means that our flake8 output will not be translated automatically
in annotations for the run.

In addition, I've split up the workflow into two separate files: one for
linting & testing and one for building (& deploying).

Signed-off-by: Sebastiaan Zeeff <sebastiaan.zeeff@gmail.com>