Fix the deploy stage of our build pipeline
I've fixed the deploy stage of our build pipeline, as it got mixed in
with the old workflow file due to a merge conflict. The deploy stage is
currently split into a separate workflow; theoretically, this allows us
to trigger a redeploy from GitHub, without having to build the container
image again.

Signed-off-by: Sebastiaan Zeeff <sebastiaan.zeeff@gmail.com>