Help channels: don't check if task is done before awaiting
Awaiting a done task is effectively a no-op, so it's redundant to check
if the task is done before awaiting it.

Furthermore, a task is also considered done if it was cancelled or an
exception was raised. Therefore, avoiding awaiting in such cases doesn't
allow the errors to be propagated and incorrectly allows the awaiter to
keep executing.