Make sure PR build artifact is always uploaded
GitHub Actions has an implicit status condition, `success()`, that
is added whenever an `if` condition lacks a status function check
of its own. In this case, while the upload step did check for the
outcome of the previous "always" step, it did not have an actual
status check and, thus, only ran on success.

Since we always want to upload the artifact, even if other steps
failed, I've added the "always" status function now.