Bot: cease support of Bot.clear()
Supporting the function means supporting re-use of a closed Bot.
However, this functionality is not relied upon by anything nor will it
ever be in the foreseeable future.

Support of it required scheduling any needed startup coroutines as
tasks. This made augmenting the Bot clunky and didn't make it easy to
wait for startup coroutines to complete before logging in.