Simplify the _split_parameters implementation
The main simplification was getting rid of keeping track of string depth
which was unnecessary, as we can just always skip them as was being
done for strings inside of brackets.

The branching was also simplified to make sure less unnecessary checks
were being done with a bit less confusing elifs.