Reviewer no longer subclasses Scheduler
It didn't make much sense for the Reviewer to subclasses Scheduler.
The Scheduler has methods that don't make sense to use on the Reviewer directly.

There is now a Scheduler object as an attribute of the Reviewer.
Interacting with it is done by adding __contains__, cancel, and cancel_all methods.