Branding: cache all available events
This is a prequel to adding a calendar command. To avoid re-querying
the branding repo on command invocation, event information will be
cached whenever we make requests. The command can then simply get an
up-to-date event schedule from the cache, with the option of forcing
an update via the 'populate_cache_events' function.

Since we cannot easily serialize entire 'Event' instances, we simply
store what's needed - the event name, and its duration.

The author has verified that the cache maintains order; in this case
chronological order based on event start date.