Branding: adjust daemon start-up behaviour
The daemon will now perform a sync iteration immediately when started,
and then every UTC midnight.

Previously, it would only perform the initial iteration when started
for the first time, which is odd.

It is also believed that splitting the daemon's logic into three
separate functions is beneficial: before, loop, and main.

This commit makes log and doc adjustments where appropriate.