Branding: raise on non-200 responses
The fetch helpers will now raise when the request fails rather than
logging a warning and returning a fallback value.

This allows better error logging as the caller is able to log the
propagated exception while adding its own context.

Additionally, the caller in some cases no longer needs to check for
the None return and raise its own exception.