Refactors code.
Earlier on message edit the message wasn't run through extract message links to see if new message links are added or if some got deleted. Similarly the cache was updated when a message got deleted.

Now it makes extract message links a helper function and runs it on message edits and deletes in case there are some changes in the message links.

This commit also updates the doc strings for functions according to the new changes done.