Fix truncation bug
When you take a long message, just one word of 400 A's then the truncated wouldn't be able to handle it properly and just return the placeholder. This is a bug in the textwrap.shorten function.

To solve this, I went the long way to use slicing on the list.

This commit seems to have resolved the bug.