Add new cog for extracting data from metabase
Metabase generates report from site and metricity data. Quite often we export
these reports to csv, transform them and the pipe into an int e. This cog aims to
reduce the time taken for that, by giving admins the ability to export data from
a report directly into a hastebin.

The auth flow is cached, as the login endpoint is ratelimitted. We want to ensure
that we use a valid session token until it expires to reduce the number of calls
to this endpoint.