Revert changes to int e
This commit reverts the changes to int e, by giving the invoker instructions
on how to access the export via int e. This means the metabase exports are not
inserted to every int e envrionment.

I have also added a seperate message in this commit to handle when the paste service
is offline.