Prefer using the package name as a prefix when handling symbol conflicts
When renaming, having label.symbol etc. in the list of the renamed symbol
in the embed doesn't really provide much context on what that symbol may be
to the user. Prioritizing the package name instead should makes it clearer that
it's from an another package.
Conflicts within a package are still resolved using the previous logic