Prioritize symbols depending on their group's pos in FORCE_PREFIX_GROUPS
When both symbols are in FORCE_PREFIX_GROUPS, instead of relying on which symbol
comes later and always renaming the latter one, the symbols with a lower
priority group are moved out of the way instead of forcing the new symbol to be
moved.
This will help make relevant symbols be more likely to come up when
searching the docs.
The constant was reordered by the priority of the groups to work with this
change