Add an optional loop kwarg to scheduling.create_task
Before this change, the create_task util couldn't be used to schedule tasks from
the init of cogs, as it relied on asyncio.create_task that uses the currently
running loop to create the task.
The loop kwarg allows the caller to pass the loop itself if there's no running
loop yet.