Time: remove timedelta and relativedelta support from discord_timestamp
When a delta is given, it is unknown what it's relative to. The function
has to assume it's relative to the POSIX Epoch. However, using a delta
for this would be quite odd, and would more likely be a mistake if
anything.

relativedelta support was broken anyway since it wasn't using the
total seconds represented by the delta.