fix: Nomination message now checks historic and new style nominations
Previously nomination messages had role pings in them, now they don't as we moved them into a thread.
Due to this, we need to detect both in the interim of historic nominations existing.

A 'proper' fix for this is to store the nomination message IDs when we post them against the
nomination object in the site api. We are planing to work on this soon, this commit is a
short term fix.