HelpChannels: fix incomplete init of available channel set
If the cog is reloaded while there are less than the maximum amount of
available channels, it makes some channels available until the limit is
reached. When a channel is made available, it updates the
`available_help_channels` set. The `update_available_help_channels()`
function would not update this set if it saw that the set already
contains elements. This resulted in only the channels that were just
made available being in the set; the set would not contain the channels
that were already available when the bot started.

Fix this by unconditionally populating the set, but moving it to
`init_available()` so it only happens once.

Fix BOT-Z1
Fix #1715