Handle message not found specially
Originally message not found would be passed into discord.Exceptions and
then would be logged which was unnecessary for a `MessageNotFound`
error.

Now this has been handled, if the bot receives a deleted message
it would look through the last 100 messages of mod_logs channel to check
if a log entry exists for that message. If one exists, then the incident
embed would have this log entry message linked to it, if it doesn't it
would send a message not found error embed.

In future, we could maybe do the modlog deleted message entry finding
via the python discord API, rather than using hard coded values and
checking if they are existing in the log entry.