Only check URL-like objects against domain filters
Previously a message such as 'https://google.com hello! flask.request.method' would be
filtered due to us filtering the url shortener t.me. This commit changes to logic so
that we only check parts of the messages that matched the URL regex against our
blacklist, to avoid these false-positives.