Change cache usage
The cache is used only when all channels are used, as before.
Unlike before, using all channels requires using "*" in the channels argument. Before all channels would be used if use_cache was set to True.
Using all channels uses the cache by default. To traverse every single text channel in the server, setting use_cache to False is required in the command.