Rename "amount" argument to "traverse"
This name as been confusing moderators for a while now.
"amount" sounds like this is the amount of messages that the bot will try to delete, and keep going until it reaches that number.
In reality it's the amount of latest messages per channel the bot will traverse.
Hopefully the new name conveys that better.