Added master command
The subcommands are kept simple and with few arguments, as they deal with most cases and their usage shouldn't be cumbersome.
However we might to clean by criteria of several functionalities offered by the subcommands, for example delete a specific user's messages but only those that contain a certain pattern.
For this reason the top-level command can now accept all arguments available in any of the subcommands, and will combine the criteria.

Because the channels list has to be the last argument in order to accept either a list of channel or "*", I had to force a specific pattern in the regex argument for it to not consume the first channel specified.
The regex argument must now have an "r" prefix and be enclosed in single quotes. For example: r'\d+'. For patterns with spaces the whole thing still needs to be enclosed in double quotes.
For consistency the `clean regex` subcommand was changed to require the same.