Prevent erroneous symbols from always raising stale warnings
Some doc symbols are improperly generated and never exist on the doc page
the inventory file defines them in, causing the stale warning to get raised
every time the page is parsed (at a maximum every week because of the redis
expire).

This can be prevented by keeping a counter in redis for the items which were
stale, every time the item is warned for the counter is incremented and set to
expire in 3 weeks.
Then a warning is only raised when the counter is below 3, resulting in the
unpreventable warning only being raised twice until it is fixed by the
maintainers after it expires in 3 weeks after the last increment.