Add concurrency rules to all GitHub workflows
This concurrency rule means that workflow actions will be cancelled if a new set of actions are triggered on the same ref. A ref is either a PR or a branch itself.

Doing this means that if someone pushes to a PR multiple times in quick succession, workflows won't get queued up. Instead, only the workflows from the most recent changes will run, as this is all we really care about anyway.

I see the benefits of this as twofold.
1. The author gets faster feedback about the most recent change pushed, rather than have to wait for all previous runs to completed
2. Other contributors don't need to wait for a queue of redundant workflows to finish.