Declare and refresh TalentPool.cache on init of cog
This avoids issues in the server cog trying to access it before it's assigned and refreshed.

I also migrated to the tasks to `scheduling.create_task()` as the created tasks currently don't have any error handling they can hide errors in development until the task object is destroyed (if that occurs at all) which logs the exception. The scheduling alternative attaches a callback which logs exceptions to prevent this.