Raise InvalidHeaderError if inventory has invalid header
If the requested inventory file has an invalid header, the resource
has most likely moved or the url was invalid. In case of adding a new
inventory, this error can be handled and displayed to the user.
For automatic fetching it'll stop the rescheduling as the request succeeded
but returned invalid content.