Add metricity to docker-compose
By adding metricity to the compose, we allow it to migrate itself, rather than needing the site to do it.

Defaulting 'USE_METRICITY' to false means that it will run migrations, but not actually start the bot. This means we don't add another service that needs to run all the time, which could impact some contribs on lower powered hardware.