Allow helpers to edit their own nomination reason
This change will allow helpers to run the edit reason command in the Talentpool cog.

To ensure that non-mod helpers can only edit their own reasons the interface for the reason command has been changed slightly.

If nominee_or_nomination_id is a member or user, then the command edits the currently active nomination for that person.
If it's an int, then use this to look up that nomination ID to edit.

If no nominator is specified, assume the invoker is editing their own nomination reason.
Otherwise, edit the reason from that specific nominator.

Raise a permission error if a non-mod staff member invokes this command on a
specific nomination ID, or with an nominator other than themselves.