Direct users to the appeals server when banned
This is a new appeals process we are trialing. Users who get banned join this server and DM a modmail bot, who relays the message to the main server for mods to discuss.

I have updated the shortening logic to allow for extra information to be included at the end of the embed, while still staying under the limit.