Dynamic views for command help embeds
Adds views for commands to navigate groups.
For subcommands, a button is added to show the parent's help embed.
For groups, buttons are added for each subcommand to show their help embeds.

The views are not generated when help is invoked in the context of an error.

The code has gone through several iterations of design, and this is what I arrived at.
Because conversion of callbacks to buttons is done in __subclass_init__ of the view baseclass, I had to create Button subclasses.
Because the normalization of button positions is in the view init and is obfuscated, I had to meddle with discord.py internals a bit in GroupView to make it work.