Deal with activity_blocks not being returned by site
We are planning to change metricity endpoints on site so that activcity_blocks are not returned if the user has more than 1000 messages. This is because the query to calculate those blocks can get expensive at a high message count.

To deal with this, both places activity_blocks are used has been changed to reflect this planned behaviour.