Redirect subscribe command output to bot commands
Instead of silently failing in channels other than bot commands for non-staff, the bot now instead redirects the command output to bot commands and pings the user.

To facilitate this, I had to change the ctx.reply to a ctx.send since the invocation message may be in a different channel.