Simplify cache usage
Removes the cache usage argument from the clean commands.
Cache usage is now an implementation detail.
The cache will be used if the age of the oldest message requested for cleaning is younger than the oldest message in the cache.

Additionally fixes the logger to the one used in the rest of the bot (caused by a faulty merge).