Emphasize reason for sharing traceback
The previous version talks about how to read a traceback, but this tag is most often used to help the asker provide the traceback. This version is more brief and emphasizes why they are being asked to provide the traceback, and uses an example that I think is slightly more illustrative.