Add unsubscribe alias to subscribe command
In quite a few places, such as #roles, we tell users to run the unsubscribe command to remove roles from them. However, this command no longer exists due to the rework of the subscribe command.

Since the subscribe commands allows tyou to remove as well as add roles, I have added this as an alias.